multiple swaggers on the same tomcat
Swagger has somewhat global presence in tomcat environment. If you want to partition your api, global nature hiders you. This commit aims to alleviate the issue.

ScannerFactory does not act in accord with its name since this factory does not produce new instances. The class is used as the holder for Scanner instance. As I understand, there exist only one instance of Scanner by design so I find accurate embedding this fact in classes' name, hence ScannerSingleton.
For each part of my api I am gonna need proper SwaggerConfig. For instance to get proper baseUrl. Each instance of SwaggerConfig has to be easily identified in global scope. I could use ServletContext but I loose controll over object type and it would not be consistent with the way global presence of Scanner is done. Added SwaggerConfigMap.
It seems that ServletContext was used to store "reader" (that is obviously instance of SwaggerConfig) and "scanner". When "reader" is used, "scanner" seems to be abbandoned in favor of ScannerFactory. Moreover the "reader" actually sets global scanner. As a result "scanner" has no merit.
Since ScannerFactory is now ScannerSingleton I moved creation of new instance of Scanner from WebXMLReader to ScannerSingleton. BeanConfig replace this instance as it have done.
JerseyJaxrsConfig is present only due to compatibility (I guess) so I modified the class to reflect that.
ApiListingResource stores instances of Swagger using swagger.id what allows partitioning. When swagger.id is absent Swagger is stored under name "swagger-default" as a result previous functionality is supported.