Properly filter subclass properties
When a class is a referred to in a @ApiModel(subClasses) the
properties from the base class are supposed to be removed from
the subclass. There was code to do this, but the properties of
the base class had not been set before an attempt was made to
loop through them.

Added test to verify behavior