issue #1667
Removal of `Property#setExample(String)` can cause linkage errors at runtime, if there are conflicting versions of swagger-models. For example, springfox-swagger2 uses swagger-core:1.5.6, which pulls in swagger-models:1.5.6. If swagger-codegen is also in the classpath, `Property#setExample(String)` from the swagger-models:1.5.6 clashes with `Property#setExample(Object)`; overriding the transitive dependency is not an option due to the incompatible signature-change. This change adds the method back to the `Property` class but marks it as deprecated, thereby preserving backwards-compatibility.