Look for parameter annotations on parent classes/interfaces.
This is a fix for issue #1506.
Reader got parameter annotations from a child method, ignoring any
on a parent class or interface. This fix will look for annotation
on the parent class or interface, if it exists. Credit mohitmutha
with the fix. I just implemented it and created the pull request.