Updated ModelResolver to generate parent model and other subtype models if a child type meets certain conditions
Update ModelResolver so that if a child type is resolved and meets certain conditions, then a model will be generated for its parent type, and all other sub-types of that parent type. This ensures that the inheritance relationship between the child type and its parent type is correctly captured in the swagger schema as a parent model and a child composition model with a reference to the parent model. Previously, this was captured as a single child model containing the parent properties. This will continue to be the case unless the specific conditions are met.

The conditions under which the above will occur include the following:
1. The child type must have an ApiModel annotation specifying its parent type in the parent property of the annotation.
2. The child type must in fact be a child of the parent type, as determined by the AnnotationIntrospector instance (_intr) in the parent AbstractModelConverter class.
3. The ApiModel annotation in the parent class must specify a subTypes property containing the child type.

Creating a parent model and children composition models with references to the parent model already occurs when a parent type is resolved using the ModelResolver. By updating the ModelResolver to create a similar set of classes when resolving a child type (under the aforementioned conditions), the ModelResolver is consistent in its model generation and provides greater clarity regarding inheritance relationships.