Handle complex types in the schema.
A type declaration may be an array of strings, as opposed to a single string.
An object matches this schema if it matches any one of the named types.

Handle this by mapping the type schema to an appropriate simple type for
the property to use, e.g. by mapping ["string", "null"] to "string" and
["string", "bool"] to "object".