[osgi] Update to a more recent maven-bundle-plugin
Swagger was using an ancient (mid 2012) version of the maven-bundle-plugin. This uses an even older version of bnd to work out what Swagger's OSGi metadata should be. This old version of bnd uses a naive algorithm for calculating uses constraints which often overspecifies, in this case putting constraints on Jackson and Guava which are not needed. This in turn makes Swagger much harder to deploy in OSGi as Guava's versioning policy is not semantic.

Switching to a more modern maven-bundle plugin fixes this.

Signed-off-by: Tim Ward <timothyjward@apache.org>