Response Parsing Order Changes
Responses are currently pared in the order of: defaults->method level->class level.
This order means that class level responses have to make sure not to overwrite method level responses when they are added to the Operation object.

With the current implementation class level default responses (eg. 200) cannot override default responses.
The order has been changed to defaults->class level->method level to prevent this issue and to simplify adding responses to the Operation object.