ref #3133 - Add swagger config file option to maven plugin
Currently the maven plugin only accepted OpenAPI files. This adds
the option to provide a swagger configuration (e.g. prettyPrint),
so if you already have a config file for say, serving a life api doc,
you don't have to define it twice.

This heavily refactors the maven plugin and adds a bunch of tests. Also
the readme is extended to reflect the new parameter. The error handling
for file reading was changed to fail-fast, meaning, if a user provides
the config, the file must exist and must be readable, otherwise an error
is thrown. This makes it easier to correctly set the plugin up.