fix: use arraySchema when deciding required of an array (#4998) Fixes: #4341
ArraySchema.schema describes the schema of the items.
ArraySchema.arraySchema describes the schema of the array itself (i.e., the property).

This is the same conceptual bug as #3438, which concerned the old boolean required attribute on @Schema. That older issue was fixed by switching from arraySchema.schema().required() to arraySchema.arraySchema().required().